"For Microsoft Dynamics AX 2012 R2..."
This section needs to be changed - "For Microsoft Dynamics AX 2012 R2, use the version of the Data Import/Export Framework that is available in cumulative update 7 for Microsoft Dynamics AX 2012 R2." 
DIEF in CU7 is a SYP model.  So it only applies to AX2012R2 CU7 or higher CU in R2 environment.  DIEF in CU7 just installs the binary (dlls and config files) it doesn't change the the models. So, for a customer on AX2012 R2 CU1, asking them to install DIEF from the CU7 binary package is not going to work. I just verified DIEF in Cu7 is a SYP model in a CU7 environment.